requirements.txt
py3nvml==0.2.5
python-dotenv==0.10.3
schedule==0.6.0
slackclient==2.5.0
SQLAlchemy==1.3.12
